SendAmountTextFieldFormatter
- Move to common App
