+++ LocalPackages/App/Sources/App/SendModule/Modules/Amount/AmountInput/AmountInputViewController.swift
@@ -38,19 +38,7 @@ final class AmountInputViewController: GenericViewViewController<AmountInputView
     }
   }
   
-  private let sendAmountTextFieldFormatter: SendAmountTextFieldFormatter = {
-    let numberFormatter = NumberFormatter()
-    numberFormatter.groupingSeparator = " "
-    numberFormatter.groupingSize = 3
-    numberFormatter.usesGroupingSeparator = true
-    numberFormatter.decimalSeparator = Locale.current.decimalSeparator
-    numberFormatter.maximumIntegerDigits = 16
-    numberFormatter.roundingMode = .down
-    let amountInputFormatController = SendAmountTextFieldFormatter(
-      currencyFormatter: numberFormatter
-    )
-    return amountInputFormatController
-  }()
+  private let sendAmountTextFieldFormatter = SendAmountTextFieldFormatterFactory.make()
   
   override func viewDidLoad() {
     super.viewDidLoad()
