+++ LocalPackages/App/Sources/App/SendModule/Modules/SendV3/SendV3ViewModel.swift
@@ -221,19 +221,7 @@ final class SendV3ViewModelImplementation: SendV3ViewModel, SendV3ModuleOutput, 
   
   // MARK: - Formatters
   
-  let sendAmountTextFieldFormatter: SendAmountTextFieldFormatter = {
-    let numberFormatter = NumberFormatter()
-    numberFormatter.groupingSeparator = " "
-    numberFormatter.groupingSize = 3
-    numberFormatter.usesGroupingSeparator = true
-    numberFormatter.decimalSeparator = Locale.current.decimalSeparator
-    numberFormatter.maximumIntegerDigits = 16
-    numberFormatter.roundingMode = .down
-    let amountInputFormatController = SendAmountTextFieldFormatter(
-      currencyFormatter: numberFormatter
-    )
-    return amountInputFormatController
-  }()
+  let sendAmountTextFieldFormatter = SendAmountTextFieldFormatterFactory.make()
   
   // MARK: - Dependencies
